WIP time limit, working but need to clone game is I don't need to roll back

+++ src/utils.rs
@@ -1,8 +1,8 @@
 
 use std::time::{SystemTime, UNIX_EPOCH};
 
-pub fn get_time() -> i32 {
+pub fn get_time() -> u128 {
     let start = SystemTime::now();
     let since_the_epoch = start.duration_since(UNIX_EPOCH).expect("Time went backwards");
-    return since_the_epoch.as_millis() as i32;
+    return since_the_epoch.as_millis();
 }
